+++ src/test/kotlin/no/nav/helse/modell/person/HentInfotrygdutbetalingerLøsningTest.kt
@@ -24,22 +24,22 @@ internal class HentInfotrygdutbetalingerLøsningTest {
 
     @Test
     fun `oppdater infotrygdutbetalinger til person`() {
-        every { dao.findInfotrygdutbetalinger(FNR.toLong()) } returns "{}"
+        every { dao.findInfotrygdutbetalinger(FNR) } returns "{}"
         val json = objectMapper.createObjectNode()
         val utbetalinger = HentInfotrygdutbetalingerLøsning(json)
         utbetalinger.oppdater(dao, FNR)
-        verify(exactly = 1) { dao.updateInfotrygdutbetalinger(FNR.toLong(), json) }
+        verify(exactly = 1) { dao.updateInfotrygdutbetalinger(FNR, json) }
     }
 
     @Test
     fun `lagre ny infotrygdutbetalinger når person ikke har fra før`() {
         val utbetalingerRefId = 1
-        every { dao.findInfotrygdutbetalinger(FNR.toLong()) } returns null
+        every { dao.findInfotrygdutbetalinger(FNR) } returns null
         every { dao.insertInfotrygdutbetalinger(any()) } returns utbetalingerRefId
         val json = objectMapper.createObjectNode()
         val utbetalinger = HentInfotrygdutbetalingerLøsning(json)
         utbetalinger.oppdater(dao, FNR)
         verify(exactly = 1) { dao.insertInfotrygdutbetalinger(json) }
-        verify(exactly = 1) { dao.updateInfotrygdutbetalingerRef(FNR.toLong(), utbetalingerRefId) }
+        verify(exactly = 1) { dao.updateInfotrygdutbetalingerRef(FNR, utbetalingerRefId) }
     }
 }
